+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Update savehar addon to handle scenarios where "path" key in cookie
   attrs dict is missing.
   ([#6458](https://github.com/mitmproxy/mitmproxy/pull/6458), @pogzyb)
-* Add `server_replay_404_extra` option to serverplayback to return 404
+* Add `server_replay_extra` option to serverplayback to define behaviour
   when replayable response is missing.
   ([#6465](https://github.com/mitmproxy/mitmproxy/pull/6465), @dkarandikar)